Revert "Feat/rotate game 90 degrees"
