+++ .gitignore
@@ -1,65 +1,2 @@
-# macOS
-.DS_Store
-.AppleDouble
-.LSOverride
-
-# Thumbnails
-._*
-
-# Files that might appear in the root
-elata-eeg/
-
-# Node
-node_modules/
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-package-lock.json
-pnpm-lock.yaml
-
-# build outputs
-dist/
-build/
-out/
-/.cache/
-
-# Logs
-*.log
-
-# IDEs and editors
-.vscode/
-.idea/
-*.suo
-*.ntvs*
-*.njsproj
-*.sln
-*.swp
-
-# Python
-__pycache__/
-*.py[cod]
-*.egg-info/
-dist-python/
-
-# OS temporary files
-Thumbs.db
-ehthumbs.db
-Desktop.ini
-
-# Tone/Vendor or other large vendors if you prefer to keep them out of source
-Pongo/vendor/
-
-# Misc
-.env
-.env.local
-.env.*.local
-
-# Ignore archives and useless binaries
-*.zip
-*.tar.gz
-*.gzip
-*.rar
-*.7z
-
 elata-eeg/
