+++ Pongo/server/package.json
@@ -5,7 +5,6 @@
   "type": "commonjs",
   "scripts": {
     "start": "node server.js",
-    "dev": "node --watch server.js",
     "test": "jest"
   },
   "dependencies": {
