--- docs/how-to-play-pong.md
@@ -1,63 +1,0 @@
-# How to Play Pongo
-
-Welcome to Pongo - the classic Pong game you can play online with friends!
-
-## 🎮 Getting Started
-
-1. **Click "Play Now"** from the main page
-2. **Share the game URL** with a friend to play together
-3. **Use your keyboard** to control your paddle
-
-## 🕹️ Controls
-
-### Player 1 (Bottom Paddle)
-- **A** = Move paddle left
-- **D** = Move paddle right
-
-### Player 2 (Top Paddle)  
-- **← Arrow Key** = Move paddle left
-- **→ Arrow Key** = Move paddle right
-
-## 🎯 How to Win
-
-- **Hit the ball** with your paddle to send it back to your opponent
-- **Score points** when the ball gets past your opponent's paddle (exits top/bottom)
-- **First to reach the target score wins!**
-
-## 🎨 Game Features
-
-- **Smooth gameplay** - No lag or delays
-- **Beautiful effects** - Neon glow and sound effects
-- **Progressive difficulty** - Ball gets faster during long rallies
-- **Fair play** - Server ensures everything stays synchronized
-
-## 🤝 Playing with Friends
-
-1. **Start a game** by clicking "Play Now"
-2. **Copy the URL** from your browser's address bar
-3. **Send the URL** to your friend
-4. **They join** by opening the same URL
-5. **Game starts automatically** when both players are ready!
-
-## 💡 Tips
-
-- **Stay calm** - The ball speeds up during long rallies
-- **Watch the angle** - Hit the ball with different parts of your paddle to change direction
-- **Practice timing** - Good timing beats button mashing
-- **Have fun!** - It's all about enjoying the game with friends
-
-## 🔧 Troubleshooting
-
-**Game not working?**
-- Make sure you have a stable internet connection
-- Try refreshing the page
-- Check that JavaScript is enabled in your browser
-
-**Can't connect with a friend?**
-- Make sure you're both using the same game URL
-- Wait a moment for the connection to establish
-- Try refreshing both browsers
-
----
-
-*Need more help? Join our [community Discord](https://discord.com/invite/GqS9CstffK) for support!*
