--- public/apps/pongo/style.css
@@ -1,118 +1,0 @@
-body {
-    font-family: 'Press Start 2P', cursive;
-    background: #0f0c29;
-    background: -webkit-linear-gradient(to right, #24243e, #302b63, #0f0c29);
-    background: linear-gradient(to right, #24243e, #302b63, #0f0c29);
-    color: #ffffff;
-    overflow: hidden; /* Prevent scrolling */
-}
-canvas {
-    background-color: rgba(0, 0, 0, 0.5);
-    display: block;
-    border: 2px solid #00ffff;
-    box-shadow: 0 0 25px #00ffff, inset 0 0 15px rgba(0, 255, 255, 0.5);
-    border-radius: 10px;
-    cursor: none; /* Hide cursor over canvas */
-}
-.title, .score-label {
-     color: #00ffff;
-     text-shadow: 0 0 10px #00ffff, 0 0 20px #00ffff;
-}
-.score {
-    color: #ffffff;
-    text-shadow: 0 0 10px #00ffff, 0 0 20px #00ffff, 0 0 30px #00ffff;
-}
-.message-box {
-    background-color: rgba(0, 0, 0, 0.6);
-    border: 2px solid #00ffff;
-    box-shadow: 0 0 20px rgba(0, 255, 255, 0.7);
-    backdrop-filter: blur(5px);
-}
-.restart-btn {
-    background-color: #00ffff;
-    color: #0f0c29;
-    text-shadow: none;
-    box-shadow: 0 0 15px #00ffff;
-    transition: all 0.3s ease;
-}
-.restart-btn:hover {
-    background-color: #ffffff;
-    color: #000;
-    box-shadow: 0 0 25px #ffffff;
-}
-#muteButton {
-    text-shadow: 0 0 10px #00ffff;
-}
-/* Minimal utility replacements so UI works without Tailwind */
-.hidden { display: none !important; }
-#startOverlay {
-    position: fixed;
-    inset: 0;
-    display: flex; /* Show by default */
-    align-items: center;
-    justify-content: center;
-    z-index: 50;
-    background-color: rgba(0, 0, 0, 0.7);
-}
-#messageBox, #speedModal {
-    position: fixed;
-    inset: 0;
-    display: none; /* Hidden by default */
-    align-items: center;
-    justify-content: center;
-    z-index: 50;
-}
-#messageBox { z-index: 60; }
-#countdown { display: none; }
-.toast {
-    position: fixed;
-    top: 20px;
-    right: 20px;
-    background-color: rgba(0, 255, 255, 0.9);
-    color: #0f0c29;
-    padding: 12px 24px;
-    border-radius: 8px;
-    box-shadow: 0 0 20px rgba(0, 255, 255, 0.7);
-    font-family: 'Press Start 2P', cursive;
-    font-size: 12px;
-    z-index: 1000;
-    opacity: 0;
-    transform: translateY(-20px);
-    transition: all 0.3s ease;
-}
-.toast.show {
-    opacity: 1;
-    transform: translateY(0);
-}
-
-
-/* Minimal utility classes to replace Tailwind used in HTML */
-.flex { display: flex; }
-.flex-col { flex-direction: column; }
-.items-center { align-items: center; }
-.justify-center { justify-content: center; }
-.justify-start { justify-content: flex-start; }
-.justify-around { justify-content: space-around; }
-.w-full { width: 100%; }
-.max-w-4xl { max-width: 64rem; margin-left: auto; margin-right: auto; }
-.text-center { text-align: center; }
-.mb-2 { margin-bottom: 0.5rem; }
-.mb-4 { margin-bottom: 1rem; }
-.my-2 { margin-top: 0.5rem; margin-bottom: 0.5rem; }
-.p-2 { padding: 0.5rem; }
-.p-4 { padding: 1rem; }
-.ml-2 { margin-left: 0.5rem; }
-
-/* Center the whole app content and keep space-efficient on small screens */
-body { display: flex; flex-direction: column; align-items: center; justify-content: flex-start; min-height: 100vh; padding: 16px; }
-
-/* Responsive typography using clamp() */
-.title { font-size: clamp(24px, 5vw, 56px); text-align: center; }
-.score-label { font-size: clamp(14px, 2.6vw, 28px); }
-.score { font-size: clamp(24px, 6vw, 64px); }
-
-/* Center the scoreboard row and controls */
-.text-center { text-align: center; }
-
-/* Buttons sizing that adapts */
-.restart-btn { font-size: clamp(12px, 2.2vw, 14px); padding: 0.5rem 0.75rem; border-radius: 0.5rem; }
